Put in a ksh fix for the configure.in Make PILS into a subpackage. Install header files in different directories for PILS and STONITH. More IGNORESIG fixes.
CVS patchset: 1222
CVS date: 2002/08/10 02:26:12

+++ telecom/apphbd/apphbtest.c
@@ -72,7 +72,7 @@ doatest(void)
 	int	rc;
 
 	sleep(5);
-	fprintf(stderr, "Client starting - pid: %d\n", (int) getpid());
+	fprintf(stderr, "Client starting - pid: %ld\n", (long) getpid());
 	rc = apphb_register("test program");
 	if (rc < 0) {
 		perror("registration failure");
